Refactor server routes into modular routers
